requirements.txt
google-generativeai==0.8.3
typing==3.7.4.3
pytest==8.3.3
logging==0.4.9.6
